--- src/model/admin.js
@@ -1,11 +1,0 @@
-class Admin {
-    constructor(id, nome, sobrenome, email, senha){
-        this.id = id,
-        this.nome = nome,
-        this.sobrenome = sobrenome,
-        this.email = email,
-        this.senha = senha
-    }
-}
-
-export default Admin;
