+++ src/models/usuario.js
@@ -1,0 +1,10 @@
+class Usuario {
+    constructor( nome, sobrenome, email, senha){
+        this.nome = nome,
+        this.sobrenome = sobrenome,
+        this.email = email,
+        this.senha = senha
+    }
+}
+
+export default Usuario;
